Pickup new buildpack version from go.mod
This will help ensure we are testing the same version of a buildpack
in the go mod and the builder. Moreover, since dependabot will send PRs
for go.mod updates, we can not automatically test the release buildpack
of the same version.

Adding golang references to the buildpacks as a faux dependency so they
will also be managed by dependabot.

The templated-template is a bit odd, but currently necessary as we want
to resolve and pin function buildpack snapshots, while the cf buildpacks
are tagged releases that we can pickup safely since the version is
committed as part of go.mod